+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [67.5.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@67.5.2...@codecademy/gamut@67.5.3) (2025-12-05)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+### [67.5.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@67.5.1...@codecademy/gamut@67.5.2) (2025-12-04)
+
+### Bug Fixes
+
+- **InfoTip:** wrap focus ([77457a2](https://github.com/Codecademy/gamut/commit/77457a2b918eea5601b77e74ba2517fc22484a7c))
+
 ### [67.5.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@67.5.0...@codecademy/gamut@67.5.1) (2025-12-03)
 
 **Note:** Version bump only for package @codecademy/gamut